refactor(afk): change delay to ticks instead of seconds

+++ core/src/main/java/net/flectone/pulse/config/Message.java
@@ -106,7 +106,7 @@ public record Message(
             Boolean enable,
             Boolean trackPlaytime,
             Range range,
-            Integer delay,
+            Long delay,
             Double radius,
             List<String> ignore,
             Destination destination,
